wykonanie 1 zadania

+++ wrz.20/przeddiagnoza/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System.Runtime.CompilerServices;
 
-byte w1=30, w2=40, odl=50;
-for (short i = 1; i <= odl; i++)
+sbyte w1=30, w2=40, odl=50;
+for (sbyte i = 1; i <= odl; i++)
 {
     if(w1*w1 + i*i == w2*w2 + (odl-i)*(odl-i)) 
     {
